# docs/openapi.json
{
  "components": {
    "schemas": {
      "PageableResponse": {
        "description": "Metadados da solicitação de página.",
        "properties": {
          "offset": {
            "description": "Deslocamento do primeiro registro.",
            "example": 0,
            "format": "int64",
            "type": "integer"
          },
          "pageNumber": {
            "description": "Número da página, iniciado em zero.",
            "example": 0,
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "description": "Quantidade solicitada por página.",
            "example": 20,
            "format": "int32",
            "type": "integer"
          },
          "paged": {
            "description": "Indica uma consulta paginada.",
            "example": true,
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/SortResponse",
            "description": "Ordenação solicitada."
          },
          "unpaged": {
            "description": "Indica uma consulta não paginada.",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "Erro HTTP no formato RFC 9457.",
        "properties": {
          "detail": {
            "description": "Explicação específica do erro.",
            "example": "Dados inválidos",
            "type": "string"
          },
          "instance": {
            "description": "URI da requisição que originou o erro.",
            "type": "string"
          },
          "status": {
            "description": "Código HTTP.",
            "example": 400,
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "description": "Título correspondente ao status HTTP.",
            "example": "Bad Request",
            "type": "string"
          },
          "type": {
            "description": "Identificador do tipo do problema.",
            "example": "about:blank",
            "type": "string"
          }
        },
        "type": "object"
      },
      "SortResponse": {
        "description": "Estado da ordenação da página.",
        "properties": {
          "empty": {
            "description": "Indica se não há ordenação.",
            "example": true,
            "type": "boolean"
          },
          "sorted": {
            "description": "Indica se existe ordenação.",
            "example": false,
            "type": "boolean"
          },
          "unsorted": {
            "description": "Indica se a consulta está sem ordenação.",
            "example": true,
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "SupplierPageResponse": {
        "description": "Página de fornecedores.",
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/SupplierResponse"
            },
            "type": "array"
          },
          "empty": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "number": {
            "format": "int32",
            "type": "integer"
          },
          "numberOfElements": {
            "format": "int32",
            "type": "integer"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableResponse"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "sort": {
            "$ref": "#/components/schemas/SortResponse"
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "SupplierRequest": {
        "description": "Dados para criação ou atualização de um fornecedor.",
        "properties": {
          "name": {
            "description": "Nome do fornecedor.",
            "example": "TOTVS",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "SupplierResponse": {
        "description": "Fornecedor cadastrado.",
        "properties": {
          "id": {
            "description": "Identificador do fornecedor.",
            "example": 1,
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "description": "Nome do fornecedor.",
            "example": "TOTVS",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "description": "Credenciais configuradas por APP_SECURITY_USERNAME e APP_SECURITY_PASSWORD.",
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API para gestão de fornecedores, contas a pagar, importações assíncronas em CSV\ne relatórios de pagamentos. Todos os recursos exigem autenticação HTTP Basic.\n",
    "title": "Payable Management API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/supplier": {
      "get": {
        "description": "Retorna os fornecedores em uma página com ordenação opcional.",
        "operationId": "listSuppliers",
        "parameters": [
          {
            "description": "Número da página, iniciado em zero.",
            "in": "query",
            "name": "page",
            "schema": {
              "default": 0,
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Quantidade de registros por página.",
            "in": "query",
            "name": "size",
            "schema": {
              "default": 20,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Ordenação no formato propriedade,direção. Pode ser repetido.",
            "example": "name,asc",
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierPageResponse"
                }
              }
            },
            "description": "Página de fornecedores."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          }
        },
        "summary": "Listar fornecedores",
        "tags": [
          "Fornecedores"
        ]
      },
      "post": {
        "description": "Cadastra um fornecedor com nome obrigatório.",
        "operationId": "createSupplier",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "name": "TOTVS"
              },
              "schema": {
                "$ref": "#/components/schemas/SupplierRequest"
              }
            }
          },
          "description": "Dados do novo fornecedor.",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierResponse"
                }
              }
            },
            "description": "Fornecedor criado."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Dados inválidos",
                  "status": 400,
                  "title": "Bad Request",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Nome ausente ou inválido."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          }
        },
        "summary": "Criar fornecedor",
        "tags": [
          "Fornecedores"
        ]
      }
    },
    "/supplier/{id}": {
      "delete": {
        "description": "Exclui um fornecedor que não possua contas a pagar vinculadas.",
        "operationId": "deleteSupplier",
        "parameters": [
          {
            "description": "Identificador do fornecedor.",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Fornecedor excluído."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Recurso não encontrado",
                  "status": 404,
                  "title": "Not Found",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Fornecedor não encontrado."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "O recurso possui vínculos e não pode ser excluído",
                  "status": 409,
                  "title": "Conflict",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Fornecedor possui contas vinculadas."
          }
        },
        "summary": "Excluir fornecedor",
        "tags": [
          "Fornecedores"
        ]
      },
      "get": {
        "description": "Retorna um fornecedor pelo identificador.",
        "operationId": "findSupplierById",
        "parameters": [
          {
            "description": "Identificador do fornecedor.",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierResponse"
                }
              }
            },
            "description": "Fornecedor encontrado."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Recurso não encontrado",
                  "status": 404,
                  "title": "Not Found",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Fornecedor não encontrado."
          }
        },
        "summary": "Consultar fornecedor",
        "tags": [
          "Fornecedores"
        ]
      },
      "put": {
        "description": "Atualiza o nome do fornecedor. O identificador da rota prevalece sobre eventual id do corpo.",
        "operationId": "updateSupplier",
        "parameters": [
          {
            "description": "Identificador do fornecedor.",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "name": "TOTVS"
              },
              "schema": {
                "$ref": "#/components/schemas/SupplierRequest"
              }
            }
          },
          "description": "Novos dados do fornecedor.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierResponse"
                }
              }
            },
            "description": "Fornecedor atualizado."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Dados inválidos",
                  "status": 400,
                  "title": "Bad Request",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Nome ausente ou inválido."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Recurso não encontrado",
                  "status": 404,
                  "title": "Not Found",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Fornecedor não encontrado."
          }
        },
        "summary": "Atualizar fornecedor",
        "tags": [
          "Fornecedores"
        ]
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ],
  "servers": [
    {
      "description": "Servidor da aplicação",
      "url": "/management/rest"
    }
  ],
...
}